requirements.txt
matplotlib>=3.5.0
pandas>=1.3.0
python-dateutil>=2.8.0
pyyaml>=6.0
